move all folder to src and create health check route
