+++ src/main.go
@@ -4,8 +4,8 @@ import (
 	"flag"
 	"os"
 
-	"github.com/Real-Dev-Squad/wisee-backend/routes"
-	"github.com/Real-Dev-Squad/wisee-backend/utils"
+	"github.com/Real-Dev-Squad/wisee-backend/src/routes"
+	"github.com/Real-Dev-Squad/wisee-backend/src/utils"
 )
 
 func main() {
